refactor(security): extract PasswordHasher
